# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Followup API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/comment/{comment}/replies": {
            "get": {
                "tags": [
                    "replies"
                ],
                "summary": "comment replies",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "replies"
                ],
                "summary": "add new replie",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/reply/{reply}/update": {
            "post": {
                "tags": [
                    "replies"
                ],
                "summary": "update replie",
                "parameters": [
                    {
                        "name": "reply",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/reply/{reply}/delete": {
            "post": {
                "tags": [
                    "replies"
                ],
                "summary": "delete replie",
                "parameters": [
                    {
                        "name": "reply",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
        "/requests/forms": {
            "get": {
                "tags": [
                    "requests forms"
                ],
                "summary": "Get All Forms",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "requests forms"
                ],
                "summary": "add New form",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "team_id",
                        "in": "formData",
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/requests/forms/{form}/update": {
            "post": {
                "tags": [
                    "requests forms"
                ],
                "summary": "add update form",
                "parameters": [
                    {
                        "name": "form",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "title",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
    },
...
}